requirements.txt
torch==1.11.0
torchvision==0.12.0
numpy
tqdm
tensorboard
matplotlib
pandas

speechbrain~=0.5.12
wandb
torchaudio~=0.11.0
pillow

gdown
librosa
